requirements.txt
psycopg2
requests
pyyaml
 